+++ src/js/components/game-root.jsx
@@ -1,7 +1,7 @@
 import React, { Component, Fragment } from "react";
 import { PropTypes } from "prop-types";
 
-import { Tile } from "../models/model-tile";
+import Tile from "../models/tile";
 import { Passage } from "./passage";
 import { GameHeader } from "./game-header";
 import { LevelMap } from "./level-map";
