+++ src/js/components/inventory.js
@@ -2,7 +2,7 @@ import React, { Component } from "react";
 import PropTypes from "prop-types";
 
 import inventoryStore from "../stores/inventory";
-import Item from "../models/model-item";
+import Item from "../models/item";
 import { ItemTile } from "./item-tile";
 
 import "../../css/components/inventory.scss";
