+++ src/js/lib/combat.js
@@ -21,8 +21,8 @@ import playHistoryStore from "../stores/play-history";
 import combatStore from "../stores/combat";
 import characterStore from "../stores/character";
 import inventoryStore from "../stores/inventory";
-import { Damage } from "../models/model-damage";
-import Defense from "../models/model-defense";
+import Damage from "../models/damage";
+import Defense from "../models/defense";
 import { getRandomNum } from "./util";
 import constants from "../constants";
 import combatConstants from "../constants/combat";
