+++ src/js/lib/monster-factory.js
@@ -1,7 +1,7 @@
-import { Monster } from "../models/model-monster";
-import { Treasure } from "../models/model-treasure";
-import { Damage } from "../models/model-damage";
-import OpponentAttack from "../models/model-opponent-attack";
+import Monster from "../models/monster";
+import Treasure from "../models/treasure";
+import Damage from "../models/damage";
+import OpponentAttack from "../models/opponent-attack";
 import itemsStore from "../stores/items";
 
 /**
