+++ src/js/models/armor.js
@@ -1,4 +1,34 @@
-import Item from "./model-item";
+import Item from "./item";
+
+function validateProps(props) {
+  const {
+    protection,
+    protectedAgainst,
+    vulnerableTo,
+    evasionMod,
+    accuracyMod
+  } = props;
+  if (typeof protection !== "number" || isNaN(protection)) {
+    return false;
+  }
+
+  if (!Array.isArray(protectedAgainst)) {
+    return false;
+  }
+
+  if (!Array.isArray(vulnerableTo)) {
+    return false;
+  }
+
+  if (typeof evasionMod !== "number" || isNaN(evasionMod)) {
+    return false;
+  }
+  if (typeof accuracyMod !== "number" || isNaN(accuracyMod)) {
+    return false;
+  }
+
+  return true;
+}
 
 export default class Armor extends Item {
   constructor(props) {
@@ -44,34 +74,4 @@ export default class Armor extends Item {
   getAccuracyMod() {
     return this.data.accuracyMod;
   }
-}
-
-function validateProps(props) {
-  const {
-    protection,
-    protectedAgainst,
-    vulnerableTo,
-    evasionMod,
-    accuracyMod
-  } = props;
-  if (typeof protection !== "number" || isNaN(protection)) {
-    return false;
-  }
-
-  if (!Array.isArray(protectedAgainst)) {
-    return false;
-  }
-
-  if (!Array.isArray(vulnerableTo)) {
-    return false;
-  }
-
-  if (typeof evasionMod !== "number" || isNaN(evasionMod)) {
-    return false;
-  }
-  if (typeof accuracyMod !== "number" || isNaN(accuracyMod)) {
-    return false;
-  }
-
-  return true;
 }
